@@ -4,65 +4,68 @@
   "author": "Raymond-Yip",
   "private": true,
   "engines": {
-    "node": ">=4.1",
+    "node": ">=6.10.0",
     "npm": ">=3.1 <4"
   },
   "dependencies": {
-    "fastclick": "1.0.6",
-    "lodash": "3.9.3",
-    "normalize.css": "3.0.3",
-    "react": "^15.0.1",
-    "react-dom": "^15.0.1"
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "whatwg-fetch": "^2.0.3"
   },
   "devDependencies": {
-    "assets-webpack-plugin": "^3.2.0",
-    "autoprefixer": "^6.2.3",
-    "babel": "^6.3.26",
-    "babel-cli": "^6.4.0",
-    "babel-core": "^6.4.0",
-    "babel-loader": "^6.2.1",
-    "babel-plugin-react-transform": "^2.0.0",
-    "babel-polyfill": "^6.3.14",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "babel-preset-stage-0": "^6.3.13",
-    "browser-sync": "^2.7.7",
-    "css-loader": "^0.23.0",
-    "csscomb": "^3.1.8",
+    "assets-webpack-plugin": "^3.5.1",
+    "autoprefixer": "^7.1.0",
+    "babel-cli": "^6.24.1",
+    "babel-core": "^6.24.1",
+    "babel-loader": "^7.0.0",
+    "babel-plugin-react-transform": "^2.0.2",
+    "babel-plugin-transform-react-constant-elements": "^6.23.0",
+    "babel-plugin-transform-react-inline-elements": "^6.22.0",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.5",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-polyfill": "^6.23.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "babel-register": "^6.24.1",
+    "babel-runtime": "^6.23.0",
+    "browser-sync": "^2.18.8",
+    "css-loader": "^0.28.1",
+    "csscomb": "^4.0.1",
     "del": "^1.2.0",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.8.5",
-    "gaze": "^0.5.1",
-    "html-webpack-plugin": "^2.16.0",
-    "isomorphic-style-loader": "0.0.5",
-    "jest-cli": "^0.4.18",
-    "json-loader": "^0.5.4",
-    "less": "^2.5.1",
-    "less-loader": "^2.2.2",
+    "extract-text-webpack-plugin": "^2.1.0",
+    "file-loader": "^0.11.1",
+    "html-webpack-plugin": "^2.28.0",
+    "http-proxy-middleware": "^0.17.4",
+    "less": "^2.7.2",
+    "less-loader": "^4.0.3",
     "minimist": "^1.1.1",
     "mkdirp": "^0.5.1",
     "ncp": "^2.0.0",
-    "postcss": "^5.0.13",
-    "postcss-import": "^7.1.3",
-    "postcss-loader": "^0.8.0",
+    "postcss": "^6.0.1",
+    "postcss-import": "^10.0.0",
+    "postcss-loader": "^2.0.5",
     "precss": "^1.4.0",
-    "psi": "^1.0.6",
-    "react-transform-catch-errors": "^1.0.0",
-    "react-transform-hmr": "^1.0.1",
-    "redbox-react": "^1.2.0",
-    "run-sequence": "^1.1.0",
-    "source-map-support": "0.3.1",
-    "style-loader": "^0.12.3",
-    "url-loader": "^0.5.6",
-    "webpack": "^1.12.10",
-    "webpack-hot-middleware": "^2.6.0",
-    "webpack-middleware": "^1.4.0"
+    "react-hot-loader": "^1.3.1",
+    "react-transform-catch-errors": "^1.0.2",
+    "react-transform-hmr": "^1.0.4",
+    "redbox-react": "^1.3.6",
+    "style-loader": "^0.17.0",
+    "url-loader": "^0.5.8",
+    "walk": "^2.3.9",
+    "webpack": "^2.5.1",
+    "webpack-hot-middleware": "^2.18.0",
+    "webpack-middleware": "^1.5.1"
   },
   "scripts": {
     "clean": "babel-node tools/run clean",
     "copy": "babel-node tools/run copy",
     "bundle": "babel-node tools/run bundle",
     "build": "babel-node tools/run build",
-    "start": "babel-node tools/run start"
+    "dll": "babel-node tools/run dll",
+    "start": "babel-node tools/run start",
+    "deploy": "babel-node tools/run deploy",
+    "cdn": "babel-node tools/run cdn",
+    "release": "node tools/release"
   }
 }
