@@ -1,13 +1,16 @@
 import run from './run';
+import clean from './clean';
+import copy from './copy';
+import bundle from './bundle';
 
 /**
  * Compiles the project from source files into a distributable
- * format and copies it to the output (build) folder.
+ * format and copies it to the output (dist) folder.
  */
 async function build() {
-  await run(require('./clean'));
-  await run(require('./copy'));
-  await run(require('./bundle'));
+	await run(clean);
+	await run(copy);
+	await run(bundle);
 }
 
-module.exports = build;
+export default build;
