@@ -1,78 +1,137 @@
+import fs from 'fs';
+import path from 'path';
+import run from './run';
+import clean from './clean';
+import copy from './copy';
 import Browsersync from 'browser-sync';
 import webpack from 'webpack';
-import webpackMiddleware from 'webpack-middleware';
+import webpackMiddleware from 'webpack-dev-middleware';
 import webpackHotMiddleware from 'webpack-hot-middleware';
 import webpackConfig from './webpack.config';
+import proxy from 'http-proxy-middleware';
 
-const DEBUG = !process.argv.includes('release');
+const DEBUG = !process.argv.includes('--release');
+
+function format(time) {
+  return time.toTimeString().replace(/.*(\d{2}:\d{2}:\d{2}).*/, '$1');
+}
+
+const proxyMiddleware = proxy('**', {
+  target: 'http://localhost:5000/',
+  changeOrigin: true,
+  logLevel: 'info'
+});
 
 /**
  * Launches a development web server with "live reload" functionality -
  * synchronizing URLs, interactions and code changes across multiple devices.
  */
 async function start() {
-  await require('./build')();
+  await run(clean);
+  await run(copy);
+
+  let time = format(new Date());
+  console.log(`[${time}] Starting 'build'...`);
 
   await new Promise(resolve => {
 
     // Patch the client-side bundler configurations
     // to enable Hot Module Replacement (HMR) and React Transform
-    webpackConfig.filter(x => x.target !== 'none').forEach(config => {
+    webpackConfig.entry = {
+      app: ['webpack-hot-middleware/client'].concat(webpackConfig.entry.app)
+    };
 
-      config.entry = {
-        app: ['webpack-hot-middleware/client', './src/app.js']
-      };
-      config.plugins.push(new webpack.HotModuleReplacementPlugin());
-      config.plugins.push(new webpack.NoErrorsPlugin());
+    webpackConfig.plugins.push(new webpack.HotModuleReplacementPlugin());
+    webpackConfig.plugins.push(new webpack.NoEmitOnErrorsPlugin());
 
-    });
+    webpackConfig
+      .module
+      .rules
+      .filter(x => x.loader === 'babel')
+      .forEach(x => (x.query = {
+        ...x.query,
 
+        // Wraps all React components into arbitrary transforms
+        // https://github.com/gaearon/babel-plugin-react-transform
+        plugins: [
+          ...(x.query ? x.query.plugins : []),
+          ['react-transform', {
+            transforms: [
+              {
+                transform: 'react-transform-hmr',
+                imports: ['react'],
+                locals: ['module'],
+              }, {
+                transform: 'react-transform-catch-errors',
+                imports: ['react', 'redbox-react'],
+              },
+            ],
+          },
+          ],
+        ],
+      }));
+
+    // Bundler compiler
     const bundler = webpack(webpackConfig);
     const wpMiddleware = webpackMiddleware(bundler, {
       // IMPORTANT: webpack middleware can't access config,
       // so we should provide publicPath by ourselves
-      publicPath: webpackConfig[0].output.publicPath,
+      publicPath: webpackConfig.output.publicPath,
 
       // Pretty colored output
-      stats: webpackConfig[0].stats,
+      stats: webpackConfig.stats,
 
       // For other settings see
       // https://webpack.github.io/docs/webpack-dev-middleware
     });
 
-    const hotMiddlewares = bundler
-      .compilers
-      .filter(compiler => compiler.options.target !== 'none')
-      .map(compiler => webpackHotMiddleware(compiler));
+    const hotMiddleware = webpackHotMiddleware(bundler);
 
-    const bs = Browsersync.create();
+    function routeMiddleware(req, res, next) {
+      const url = req.url;
+      const routerMap = {
+        '/': 'mobile'
+      };
+      const filename = path.join(bundler.outputPath, `../content/${routerMap[url]}.html`);
+      if (!routerMap[url]) {
+        next();
+        return;
+      }
 
-    bs.init({
-      server: {
-        baseDir: "./build",
-        routes: {
-          '/': './build/public',
-        },
-        middleware: [wpMiddleware, ...hotMiddlewares]
-      },
-      // no need to watch '*.js' here, webpack will take care of it for us,
-      // including full page reloads if HMR won't work
-      files: [
-        'build/public/**/*.css',
-        'build/html/*.html',
-      ]
-    });
-
-    bundler.plugin('done', function(stats) {
-      stats.stats.forEach((stat) => {
-        stat.compilation.children = stat.compilation.children.filter((child) => {
-          return child.name !== 'extract-text-webpack-plugin';
-        });
+      bundler.outputFileSystem.readFile(filename, (err, result) => {
+        if (err) {
+          next(err);
+          return;
+        }
+        res.setHeader('content-type', 'text/html');
+        res.end(result);
       });
+    }
+
+    let handleServerBundleComplete = ()=>{
+      const bs = Browsersync.create();
+      const middlewareArr = [wpMiddleware, hotMiddleware, routeMiddleware, proxyMiddleware];
+
+      bs.init({
+        ...(DEBUG ? {} : { notify: false, ui: false }),
+        server: {
+          baseDir: '../dist/content',
+          middleware: middlewareArr,
+        },
+        serveStatic: ['.', path.join(__dirname, '../dist/public/')],
+        open: false,
+        // no need to watch '*.js' here, webpack will take care of it for us,
+        // including full page reloads if HMR won't work
+        files: ['src/content/**/*.*'],
+      });
+    };
+
+    bundler.plugin('done', () => {
+      handleServerBundleComplete();
     });
 
   });
 
 }
 
-module.exports = start;
+export default start;
