@@ -1,10 +1,5 @@
 import webpack from 'webpack';
-import webpackConfig from './webpack.config';
-
-/**
- * Bundles JavaScript, CSS and images into one or more packages
- * ready to be used in a browser.
- */
+import webpackConfig from './webpack.dll';
 
 function bundle() {
 	return new Promise((resolve, reject) => {
@@ -12,8 +7,6 @@ function bundle() {
 			if (err) {
 				return reject(err);
 			}
-			// Output analyse file
-			// require('fs').writeFileSync(webpackConfig.output.path + '/stats.json', JSON.stringify(stats.toJson()))
 			console.log(stats.toString(webpackConfig.stats));
 			resolve();
 		});
