@@ -1,7 +1,7 @@
 import path from 'path';
 import HtmlWebpackPlugin from 'html-webpack-plugin';
-import webpack, { DefinePlugin, BannerPlugin } from 'webpack';
-import merge from 'lodash.merge';
+import webpack, { DefinePlugin } from 'webpack';
+import extend from 'extend';
 
 // Webpack plugin that emits a extraxt css.
 import ExtractTextPlugin from 'extract-text-webpack-plugin';
@@ -9,8 +9,8 @@ import ExtractTextPlugin from 'extract-text-webpack-plugin';
 // Webpack plugin that emits a json file with assets paths.
 import AssetsPlugin from 'assets-webpack-plugin';
 
-const DEBUG = !process.argv.includes('release');
-const VERBOSE = process.argv.includes('verbose');
+const DEBUG = !process.argv.includes('--release');
+const VERBOSE = process.argv.includes('--verbose');
 
 const AUTOPREFIXER_BROWSERS = [
   'Android 2.3',
@@ -27,14 +27,28 @@ const GLOBALS = {
   '__DEV__': DEBUG
 };
 
+const srcPath = path.resolve(__dirname, '../src');
+const alias = {};
+['action', 'components', 'reducers', 'styles', 'utils'].map((el, index)=>{
+  alias[el] = path.resolve(srcPath, el);
+});
+
 const config = {
+  entry: {
+    app: './src/app.js'
+  },
+
   output: {
-    publicPath: '/',
-    sourcePrefix: '  ',
+    publicPath: DEBUG ? '' : 'public/',
+    // publicPath: DEBUG ? '' : '',
+    path: path.join(__dirname, '../dist/public'),
+    filename: DEBUG ? '[name].js?[hash]' : '[name].[hash].js',
+    // This is used for require.ensure. The setup
+    // will work without but this is useful to set.
+    chunkFilename: '[id].[chunkHash].js'
   },
 
   cache: DEBUG,
-  debug: DEBUG,
   stats: {
     colors: true,
     reasons: DEBUG,
@@ -44,102 +58,154 @@ const config = {
     chunks: VERBOSE,
     chunkModules: VERBOSE,
     cached: VERBOSE,
-    cachedAssets: VERBOSE
-  },
-
-  plugins: [
-    new webpack.optimize.OccurenceOrderPlugin()
-  ],
-
-  resolve: {
-    root: path.resolve(__dirname, '../src'),
-    modulesDirectories: ['node_modules'],
-    extensions: ['', '.js', '.jsx', '.json']
-  },
-
-  module: {
-    loaders: [{
-      test: /\.less$/,
-      loader: ExtractTextPlugin.extract([
-        'css-loader?' + (DEBUG ? 'sourceMap&' : 'minimize&'),
-        'postcss-loader',
-        'less',
-      ]),
-    }, {
-      test: /\.json$/,
-      loader: 'json'
-    }, {
-      test: /\.(png|jpg|jpeg|gif|svg|woff|woff2)$/,
-      loader: 'url?limit=10000'
-    }, {
-      test: /\.jsx?$/,
-      include: [
-        path.resolve(__dirname, '../src'),
-      ],
-      loader: 'babel-loader',
-    }],
-  },
-
-  externals: {
-    'react': 'React',
-    'react-dom': 'ReactDOM'
-  },
-
-  postcss: function plugins(bundler){
-    return [
-      require('postcss-import')({addDependencyTo: bundler}),
-      require('precss')(),
-      require('autoprefixer')({ browsers: AUTOPREFIXER_BROWSERS })
-    ];
-  },
-
-};
-
-
-// Configuration for the client-side bundle (app.js)
-const appConfig = merge({}, config, {
-  entry: {
-    app: './src/app.js'
-  },
-
-  output: {
-    path: path.join(__dirname, '../build/public'),
-    filename: DEBUG ? '[name].js?[hash]' : '[name].[hash].js'
+    cachedAssets: VERBOSE,
+    children: false
   },
 
   // Choose a developer tool to enhance debugging
   // http://webpack.github.io/docs/configuration.html#devtool
-  devtool: DEBUG ? 'source-map' : false,
+  devtool: DEBUG ? 'cheap-source-map' : false,
+
+  resolve: {
+    alias: alias,
+    modules: [srcPath, 'node_modules'],
+    extensions: ['.js', '.jsx', '.json']
+  },
+
+  module: {
+    rules: [{
+      test: /\.less$/,
+      use: DEBUG ?
+        [
+          'style-loader',
+          'css-loader',
+          {
+            loader: 'postcss-loader',
+            options: {
+              plugins: function(ctx) {
+                return [
+                  require('postcss-import')({addDependencyTo: ctx.webpack}),
+                  require('precss')(),
+                  require('autoprefixer')({ browsers: AUTOPREFIXER_BROWSERS })
+                ]
+              }
+            }
+          },
+          'less-loader'
+        ]
+        : ExtractTextPlugin.extract({
+          use: [{
+            loader: 'css-loader',
+            query: {
+              minimize: true
+            }
+          }, {
+            loader: 'postcss-loader',
+            options: {
+              plugins: function(ctx) {
+                return [
+                  require('postcss-import')({addDependencyTo: ctx.webpack}),
+                  require('precss')(),
+                  require('autoprefixer')({ browsers: AUTOPREFIXER_BROWSERS })
+                ]
+              }
+            }
+          },
+          'less-loader'],
+          publicPath: '../'
+      }),
+    }, {
+      test: /\.(png|jpg|jpeg|gif|svg|woff|woff2)$/,
+      use: [{
+        loader: 'file-loader',
+        options: {
+          name: DEBUG ? '[path][name].[ext]?[hash]' : 'img/[hash].[ext]'
+        }
+      }]
+    }, {
+      test: /\.(eot|ttf|wav|mp3)$/,
+      use: [{
+        loader: 'file-loader',
+        options: {
+          name: DEBUG ? '[path][name].[ext]?[hash]' : '[hash].[ext]',
+        }
+      }]
+    }, {
+      test: /\.jsx?$/,
+      exclude: [
+        path.resolve(__dirname, "../node_modules")
+      ],
+      include: [
+        srcPath,
+      ],
+      use: [{
+        loader: 'babel-loader',
+        options: {
+          // https://github.com/babel/babel-loader#options
+          cacheDirectory: DEBUG,
+          // https://babeljs.io/docs/usage/options/
+          babelrc: false,
+          presets: [
+            'react',
+            'es2015',
+            'stage-0',
+          ],
+          // http://babeljs.io/docs/plugins/
+          plugins: [
+            'transform-runtime',
+            ...DEBUG ? [] : [
+              'transform-react-remove-prop-types',
+              'transform-react-constant-elements'
+              //http://lukecod.es/2016/03/14/react-invariant-violation-minified-exception-ios8-webpack-babel/
+              // 'transform-react-inline-elements',
+            ]
+          ]
+        }
+      }],
+    }],
+  },
 
   // https://github.com/sporto/assets-webpack-plugin
   plugins: [
-    // Create index.html
+    // Create html
     new HtmlWebpackPlugin({
-      filename: '../index.html',
-      template: 'src/html/index.html'
+      filename: '../content/index.html',
+      template: 'src/content/index.html',
+      usemin: DEBUG ? '' : '.min',
+      minify: DEBUG ? false : {
+        removeComments: true,
+        collapseWhitespace: true
+      }
     }),
-    
-    //  Extraxt Css
-    new ExtractTextPlugin("app.css"),
+
+    new webpack.ProvidePlugin({
+      React: 'react'
+    }),
+
+    // Use Dll Reference
+    new webpack.DllReferencePlugin({
+      context: '.',
+      manifest: require( '../dist/vendor/vendor-manifest'+(DEBUG ? '' : '.min')+'.json' )
+    }),
 
     // Define free variables
     new DefinePlugin(GLOBALS),
 
     // Emit a file with assets paths
     new AssetsPlugin({
-      path: path.join(__dirname, '../build'),
+      path: path.join(__dirname, '../dist'),
       filename: 'assets.js',
       prettyPrint: true,
       processOutput: x => `module.exports = ${JSON.stringify(x)};`,
     }),
 
-    // Assign the module and chunk ids by occurrence count
-    // Consistent ordering of modules required if using any hashing ([hash] or [chunkhash])
-    new webpack.optimize.OccurenceOrderPlugin(true),
     ...(DEBUG ? [] : [
 
-      // Search for equal or similar files and deduplicate them in the output
-      new webpack.optimize.DedupePlugin(),
+      //  Extraxt Css
+      new ExtractTextPlugin({
+        filename: "css/[name].[contenthash].css",
+        allChunks: true
+      }),
 
       // Minimize all JavaScript output of chunks
       new webpack.optimize.UglifyJsPlugin({
@@ -152,6 +218,7 @@ const appConfig = merge({}, config, {
       new webpack.optimize.AggressiveMergingPlugin()
     ])
   ]
-});
 
-export default [appConfig];
+};
+
+export default config;
